refactored transformer code
